Showing airport details for queries with airport code
